+++ src/LibreSpot.Desktop/Views/CustomAppearanceSection.xaml.cs
@@ -1,5 +1,3 @@
-using System.Windows.Controls;
-using System.Windows.Input;
 using UserControl = System.Windows.Controls.UserControl;
 
 namespace LibreSpot.Desktop.Views;
@@ -10,7 +8,4 @@ public partial class CustomAppearanceSection : UserControl
     {
         InitializeComponent();
     }
-
-    private void NestedScrollRegion_OnPreviewMouseWheel(object sender, MouseWheelEventArgs e) =>
-        WorkspaceViewInteraction.BubbleMouseWheel(sender, e);
 }